requirements.txt
Package                Version
---------------------- -----------
asgiref                3.8.1
Authlib                1.3.2
certifi                2024.8.30
cffi                   1.17.0
charset-normalizer     3.3.2
contourpy              1.3.0
cryptography           43.0.0
cycler                 0.12.1
defusedxml             0.8.0rc2
Django                 5.1
django-allauth         64.2.1
fonttools              4.53.1
idna                   3.8
joblib                 1.4.2
kiwisolver             1.4.5
matplotlib             3.9.2
numpy                  2.1.0
oauthlib               3.2.2
packaging              24.1
pandas                 2.2.2
pillow                 10.4.0
pip                    24.0
plotly                 5.24.0
psycopg2-binary        2.9.9
pycparser              2.22
PyJWT                  2.9.0
pyparsing              3.1.4
python-dateutil        2.9.0.post0
python-dotenv          1.0.1
python3-openid         3.2.0
pytz                   2024.1
requests               2.32.3
requests-oauthlib      2.0.0
scikit-learn           1.5.1
scipy                  1.14.1
seaborn                0.13.2
six                    1.16.0
social-auth-app-django 5.4.2
social-auth-core       4.5.4
SQLAlchemy             2.0.32
sqlparse               0.5.1
tenacity               9.0.0
threadpoolctl          3.5.0
typing_extensions      4.12.2
tzdata                 2024.1
urllib3                2.2.2
xgboost                2.1.1
